--- lab-notes/src/App.jsx
@@ -1,13 +1,0 @@
-// import { useState } from 'react'
-import './App.css'
-import Form from './main.jsx'
-import { CgGoogle } from 'react-icons/cg';
-
-
-
-const App = () => {
-return <div className = "app">
-<Form /> <CgGoogle />
-</div>
-}
-export default App
